pctasks/run/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pctasks.run"
dynamic = ["version"]
description = "Planetary Computer Tasks framework: Runner component"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Microsoft", email = "planetarycomputer@microsoft.com" },
]
keywords = [
    "Planetary",
    "STAC",
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
]
dependencies = [
    "argo-workflows>=6.3.0,<6.4",
    "azure-batch==15.0.0b2",
    "azure-keyvault-secrets>=4.0.0,<5",
    "kubernetes",
    "networkx>=2.0.0,<3",
    "pctasks.core @ {root:parent:uri}/core",
    "pctasks.client @ {root:parent:uri}/client",
    "pctasks.task @ {root:parent:uri}/task",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "pdocs",
]

[project.entry-points."pctasks.commands"]
run = "pctasks.run.cli:run_cmd"

[project.urls]
Homepage = "https://github.com/Microsoft/planetary-computer-tasks"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "pctasks/run/version.py"

[tool.hatch.build.targets.wheel]
include = [
    "/pctasks/run",
]
